auto logout

+++ src/pages/axios.js
@@ -9,9 +9,12 @@ const axios_api = axios.create({
 });
 
 function handleUnauthorized() {
-  localStorage.clear();
-  window.location.href = "/login";
+  if (window.location.pathname !== "/login") {
+    localStorage.clear();
+    window.location.href = "/login";
+  }
 }
+
 
 axios_api.interceptors.response.use(
   (response) => response,
